+++ packages/rn-observatory/ios/RNObservatoryMetadata.h
@@ -1,0 +1,4 @@
+#import <React/RCTBridgeModule.h>
+
+@interface RNObservatoryMetadata : NSObject <RCTBridgeModule>
+@end