+++ packages/rn-observatory/react-native.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  dependency: {
+    platforms: {
+      android: {
+        packageImportPath: 'import com.gaozh1024.rnobservatory.RNObservatoryPackage;',
+        packageInstance: 'new RNObservatoryPackage()',
+      },
+    },
+  },
+};